requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
azure-ai-formrecognizer==3.3.1
azure-ai-translation-document==1.0.0
azure-common==1.1.28
azure-core==1.29.5
azure-cosmos==4.5.1
azure-identity==1.13.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.18.3
azure-storage-file-datalake==12.13.2
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
cachelib==0.10.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
configobj==5.0.8
configparser==6.0.0
cryptography==41.0.7
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fitz==0.0.1.dev2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Session==0.5.0
frozenlist==1.4.1
future==0.18.3
fuzzywuzzy==0.18.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
install==1.3.5
ipykernel==6.28.0
ipython==8.20.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.0.321
langchain-community==0.0.11
langchain-core==0.1.9
langsmith==0.0.79
looseversion==1.3.0
lxml==4.9.2
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib-inline==0.1.6
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.8
packaging==23.2
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
platformdirs==4.1.0
portalocker==2.8.2
prompt-toolkit==3.0.43
prov==2.0.0
psutil==5.9.7
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyxnat==1.6
PyYAML==6.0.1
pyzmq==25.1.2
rdflib==7.0.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.11.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
tabula==1.0.5
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.4.0
tornado==6.4
tqdm==4.65.0
traitlets==5.14.1
traits==6.3.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
yarl==1.9.4
